requirements.txt
gym>=0.21.0
opencv-python>=4.5.5.62
numpy>=1.21.4
pymavlink>=2.4.8
python-dotenv>=0.19.0